Convert domainHelpers to ts
